# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Password Vault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault:
    get:
      operationId: listVaultItems
      tags: [vault]
      summary: List vault items accessible to current user
      responses:
        "200":
          description: List of vault items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultItemSummary"

    post:
      operationId: createVaultItem
      tags: [vault]
      summary: Create a new vault item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaultItemBody"
      responses:
        "201":
          description: Created vault item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultItem"
  /vault/{id}:
    get:
      operationId: getVaultItem
      tags: [vault]
      summary: Get a vault item with its entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vault item with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultItem"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateVaultItem
      tags: [vault]
      summary: Update a vault item (name, category, entries, access)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultItemBody"
      responses:
        "200":
          description: Updated vault item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultItem"

    delete:
      operationId: deleteVaultItem
      tags: [vault]
      summary: Delete a vault item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    VaultEntry:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
    VaultItemSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        accessControl:
          type: string
        entryCount:
          type: integer
        createdBy:
          type: integer
        createdByUsername:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - description
        - accessControl
        - entryCount
        - createdBy
        - createdByUsername
        - createdAt
        - updatedAt
    VaultItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        accessControl:
          type: string
        allowedUserIds:
          type: array
          items:
            type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VaultEntry"
        createdBy:
          type: integer
        createdByUsername:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - description
        - accessControl
        - allowedUserIds
        - entries
        - createdBy
        - createdByUsername
        - createdAt
        - updatedAt
    CreateVaultItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        accessControl:
          type: string
        allowedUserIds:
          type: array
          items:
            type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VaultEntry"
      required:
        - name
        - category
        - accessControl
        - entries
    UpdateVaultItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        accessControl:
          type: string
        allowedUserIds:
          type: array
          items:
            type: integer
        entries:
          type: array
          items:
            $ref: "#/components/schemas/VaultEntry"
...